Add activesupport for time handling
We already have an implicit dependency on ActiveSupport via
govuk-rubocop so this makes this dependency explicit. The intention is
for this tool to make use of ActiveSupport's wrapping of TZInfo for
concise time zone management.

I originally had planned to use TZInfo directly however we are blocked
from using the nicer features it has since version 2 because of the
ActiveSupport dependency.