Add task to Split a Smart Survey export into CSVs
This task represents the manual process that we currently use this will
be retained so that we can use this both during the transition and as a
fallback should the automatic import fail.

The purpose of this task is to take input of a CSV file that is exported
from Smart Survey (with question rows set as question ids) and convert
it into two CSVs for each party. It uses the CSV builder class for this
that is intended to be also used in the automatic export for consistency
between files.