Add process to download responses from Smart Survey
This adds a class, Runner, which determines the times an export should
run, which is from 10am the previous day to 10am the current day in
London timezone. If this class is called before this time it will raise an
exception.

This class then interfaces with the SurveyResponseFetcher class to
download the responses from Smart Survey for the given time period and
reduce this to a hash with just the fields we want in usable names.
Users who have received disqualified responses (for our survey, those
who are under 18) are also included in this export so we need to also
filter them out to ensure all responses are "completed".

This uses Faraday as the mechanism for HTTP communication. We have
limits of 180 requests a minute with them, to work with this we have a
sleep call of 0.33 seconds between requests to try delay requests. If we
are to hit an error, such as a 429, a 5xx or a timeout, we employ a
lengthier 20 second wait before retrying to give some buffer room.

It turned out to be something of a pain to get the Faraday Retry
middleware to play nice with the Faraday RaiseError middleware (which
will raise an exception on an erroneous request) which is why the config
has ended up looking a little manual. There is an option in Faraday to
automatically retry on particular status codes however that doesn't work
if you have the RaiseError middleware enabled.