Upload CSV files to S3 after downloading
This uses the CSV builder class as a means to build CSV files for
Cabinet Office and Third Party. Once these CSV files have been built
they are uploaded to S3.

The expectation is that these will be put in a bucket as part of the
https://github.com/alphagov/transfer-coronavirus-data-service which is
built and maintained by the COVID Engineering team. I'm anticipating
we'll only need to know bucket, prefix and (potentially) credentials to
adapt to this.

For an integration test we're able to get away with just verifying that
a request went to S3 without needing to specify a mock response,
naturally this may break if S3 changes their API and expects a specific
response.