Send emails to notify partners
This adds the notifications gem so that we can interface with Notify to
send email notifications when the export is complete.

It uses template ids that are stored in the Notify account for the
"Coronavirus support team" which will be used to send notifications. The
text is defined in Notify and we inject variables of the times and count
of responses.

It expects environment variables to be configured to define the
recipients for each environment.

Again, we've been lucky that we were able to get away with a very simple
response in mocking the API call in the integration test.