Handle puts and time check in SurveyResponseFetcher
This changes the previous approach where we only used puts in the Runner
class and did the time check there. This is moved so that we can have
two different runners, one for S3 uploads and the other to output CSV
files, and not have a high degree of repeated code.

This has allowed simplifying this class a bit by avoiding the block
process and allowing us to output extra information for erroneous
scenarios.

This does come at the expense that puts calls aren't all in a single
place which is a bit messier and does make it a little harder to follow
the flow.