Create a DailyReport class
This is a class that is used to store the metadata and responses of a
daily report and may be injected into other classes for usage. It
intends to take some of the responsibilities from Runner to allow this
to be reused in other contexts.

This class is somewhat modest in what it does now but I expect it to
have a few more responsibilities once we are dealing with more responses
than just completed ones (analytics are interested in disqualified and
partial responses too).