Replace splitter task with file export
This replaces the process to split a Smart Survey exported file into
multiple CSVs with one that calls the API to generate the output CSVs.

The intention with this change is to reduce the manual steps involved in
the manual process of CSV generation and to bring this into closer
alignment with the automatic process we're building so it's less of a
dramatic switch.

This was initially prompted by Smart Survey changing their export UI
which caused a small panic that we'd lost key functionality. After that
transpired to only be a temporary glitch I concluded that this was still
a better approach given the consistency and re-use it offers.