Use Smart Survey time formatting for CSV
This is a bit of an annoying change. Since we already have precedent for
sending CSVs to interested parties a change in the time format may cause
them problems, to counter this we'll continue to use the UK format of
times.

I plan to open up conversations with the partners next week to see if a
change in date formatting would cause problems as I believe it'd be
beneficial to use a standard format for data interchange.