Bye bye DailyReport, hello Report
A new, late change is a request to do reports on more custom time ranges
(some spanning multiple days, others only a couple of hours) has meant
that Daily part of the name for this class no longer makes sense. This
is groundwork before next building the means to specify the range that
can be applied.