Add means to parse times
This allows reports to be configured with the time they will be running
for. This is intended so that custom exports can be generated based on
times that maybe multiple days or could be just a couple of hours.

This is allowed through SINCE_TIME and UNTIL_TIME which can be specified
as strings that Ruby can parse as times. These can be given relative
times such as "11:00" to change the time on the expected day (previous
day for since, current day for until) or absolute times with a date to
specify a specific date and time, for example "2020-05-01 10:00".