Export data for data labs
Each day we have been manually creating a data export for data labs,
this changes the export process to also create data labs an export. I've
added a new email template for Data Labs in Notify which is referenced
by the constant in PartnerNotifier.

This does start to build up a bit of duplication across this tool, in my
next commits I plan to add a further export (analytics) which we
increase this further and plan to clean it up then.