Adapt CsvBuilder for Performance Analyst CSV generation
This changes the CsvBuilder class to reflect that the responses
collection on a Report is no longer only completed responses. The CSV
files that require completed responses now specify this as an argument.

This adds in a CSV export for a Performance Analyst which outputs all
responses and includes data fields of start_time, submission_time, status,
user_agent, client_id, and region.

I felt that the tests no longer really needed the same empty input test
since that happens implicitly and gets quite repetitive for 4 different
files.