Remove all of the S3 and Notify code
We're not going to be going down that road anymore. The team behind
the product we were planning to use
(https://github.com/alphagov/transfer-coronavirus-data-service) are too
busy to help us access it so we're going to use Google Drive as our
means to share files and notify recipients.

This is something of a quick pass at removal where some changes, notably
the readme, will likely change quite a lot in the next commits.