Minor consistency amends
I'm about to make some changes that will make both of these inconsistent
so this is being changed in advance to avoid this getting caught up in
those diffs.

This rephrases the file_export rake task to be more specific with less
word fluff, moves csv_builder to be an attr_reader to be consistent
with an upcoming DriveExport class and fixes a missing describe block
for the FileExport spec.