Use notify as a means to send notifications
Previously we were using Google Drive API as a means to send
notifications, which was great until they stopped working. I spent a few
hours trying to diagnose any clues as to why they stopped by checking
the API responses and the Google cloud console, there weren't any clues
and our support package doesn't allow us to speak to a human.

Since these notifications are of high importance to us I felt the best
course of action was to switch to Notify. Then there is at least a means
to trace any problems with the sending of email. This felt more
productive than being blocked while awaiting help on public forums.

From all my experiments we've had no problems with the uploading or
setting the permissions of files, just the process of notifying people
that the file is shared.

To reflect the concise means of calling Notify and cross-over in
responsibilities I have expanded the DriveUploader class to handle
notifying and renamed it FileDistributor to better articulate its
responsibilities.

For this I have created a single notify template that can handle the
personalisation. This is set up in the "Coronavirus support team" notify
account that can be accessed using the
2nd-line-support@digital.cabinet-office.gov.uk address.