Create a output/slack-message.txt on drive export
This creates a text file that can be used to input dynamic content into
a slack notification. This is created so that it can be used as an
output of a Concourse pipeline and used in the step to send the
notification.

This is to work with https://github.com/cloudfoundry-community/slack-notification-resource
and this file is to be used as an input of text_file. It is then
expected to be embedded into the text of a notification with the
$TEXT_FILE_CONTENT environment variable (as per the documentation).