Add hashed email and phone fields to Data Labs output
We want an anonymised means to identify when the same email and phone
have been used to submit multiple questions. This is done by applying
a SHA256 hash function to the email and phone number fields. When doing
analysis on the data these fields can be compared to identify where a
single person may be skewing the data.

In putting this in it is understood that it is valid for users to submit
multiple questions and that that this mechanism to tag an individual
question submitter is not useful if someone enters different
emails/phone numbers each time. More this is put in as an additional
tool to help make sense of the data submitted to help trends.

To keep a reasonable degree of anonymity on the hashes they are hashed
with a secret key. This is so that only owners of the secret key can
identify that a particular email address has asked a question. This does
unfortunately mean that we need yet another environment variable. I set
this as a rather generic name, SECRET_KEY, so that it could be re-used
if we need to do anything else with a secret key.