Rename ResponsePresenter to ResponseSerialiser
This reflects that this class doesn't reflect the Presenter design
pattern and instead is used as a means of serialising the data from
Smart Survey.

After documenting some of the problems we've had on GOV.UK with
"Presenter" classes I realised I was rather guilty of the same problem
myself, so have fixed it.