Deal with partial responses marked as completed
We had the Smart Survey export fail this morning (Saturday 😿):

```
3445 total responses, 2628 completed responses
rake aborted!
NoMethodError: undefined method `+' for nil:NilClass
/tmp/build/0ac90136/govuk-ask-export/lib/ask_export/csv_builder.rb:66:in `hash'
/tmp/build/0ac90136/govuk-ask-export/lib/ask_export/csv_builder.rb:22:in `block in data_labs'
/tmp/build/0ac90136/govuk-ask-export/lib/ask_export/csv_builder.rb:21:in `map'
/tmp/build/0ac90136/govuk-ask-export/lib/ask_export/csv_builder.rb:21:in `data_labs'
/tmp/build/0ac90136/govuk-ask-export/lib/ask_export/drive_export.rb:46:in `distribute_data_labs_csv'
/tmp/build/0ac90136/govuk-ask-export/lib/ask_export/drive_export.rb:17:in `call'
/tmp/build/0ac90136/govuk-ask-export/lib/ask_export/drive_export.rb:6:in `call'
/tmp/build/0ac90136/govuk-ask-export/Rakefile:13:in `block in <top (required)>'
```

It transpires that this is caused because some of the responses with a
"completed" status have null values for some of the Smart Survey
required files. Further investigation has determined that the required
fields are only enforced client side so it is possible for these not to
be submitted if a user either has JS disabled or experiences a JS error.

The approach I've taken to solve this is to manually adapt a "completed"
response into a partial one.