Pass environment when stubbing API helpers
The introduction of the warning in 97b566e revealed that the API
stubbing helper was producing inconsistent data. When stubbing for a
live API request this stub would return a draft API response which lead
to the test suite outputting warnings.

The approach I've taken to remedy this is to forward the environment
option set when stubbing the API to be passed through to the stubbed
response. This keeps these things consistent and prevents the warnings
from occurring.