Create BigQueryExport task
This introduces a new rake task `big_query_export` that allows
downloading data from Smart Survey and using that to populate a table in
Google Big Query.

This has been added to allow our performance analyst to hook up a data
studio dashboard to the database. He specified it'd be most useful to
him to have individual tables of data that represent specific dates of
exports.

We're able to authenticate to BigQuery using the same environment
variable approach that we authenticate with Google Drive. I was actually
surprised this worked as I couldn't see it documented and thought I
would need to change to have a keyfile. Hopefully this doesn't change
and break on me.

The BigQueryExport starts by checking whether a table already exists for
a particular date (denoted by a date timestamp in the form of YYYYMMDD)
and if it does it deletes it with the expectation to re-populate it,
this is a measure we have to avoid duplicate data. A new table is then
created for the date and a concurrent process is used to populate the
data (in my testing it is really fast, 1000s of records inserted over HTTP
in a second or two).

To avoid any oddities in the bigquery data this export won't run unless
a report is specified to run from 10am one day to 10am the previous day.
This is to avoid creating any tables with data spanning beyond the date
range.

In putting this together I reflected that this approach of a daily dump
to Bigquery is probably not the best way to handle this. Had I more time
and enthusiasm I think I'd have gone for a partioned table in BigQuery
and used a callback URL from Smart Survey to hit a lambda (or Google
Cloud equivalent) function that adds a record to Bigquery. This would
provide closer to a real time data analysis.

It was a total pain putting together the integration test for this, as
the API is integrated with quite closely and bypassing authentication is
really hard. I've mostly introduced these for consistency as, if we
didn't have any already, I'd probably have not bothered going through
such pain for the value they provide.