Add daily export task
This is the task that is intended to run each day by the concourse job.
Since this runs two existing tasks I was in a tricky spot when it came
to the balance of covering this with a test against re-testing stuff
that is already tested.

I took the approach of continuing the consistency of an integration
test. I cut down the assertions made compared to the individual tests to
avoid over testing.

To make it easier to re-use code I moved the stub helpers into test
Helper classes so they could be re-used.