Add concourse tasks for big query export
This adds a task to export the current day's data which can be used if
there are problems with the current days or if we want to re-run the job
to test something.

It also a test job that uses the draft Smart Survey data and writes to a
test dataset in Big Query. This is added for ad-hoc testing of the Big
Query export.