Remove question_format field
The ask service has evolved to be used in new different ways and as a
result has removed the question_format field. To reflect this the
codebase has also been changed to remove this field.

If this field is to return it would be given a new id so this code code
not be re-instated verbatim without a change to the config.