Remove most of the code
We no longer run a daily concourse process; talk directly to Google
Drive and Notify; or do an export to BigQuery. These were operations
that were used when this was a daily service that was orchestrated. Now
that we are running ad-hoc openings and closings of the service these
things aren't used. I'm removing the code for these so the project can
be simpler. It can always be pulled back out of the git history if it is
needed again.