Add pipeline class and config
This introduces a Pipeline class that represents a transformation and
how the data is exported from Smart Survey. There will be a pipeline for
each file that is currently generated. This will be used to abstract
away the different files that need to generated and exported. This is to
remove hard coding configuration in code and the tests.