Move computed methods to the report
To reduce the complexity of the CSV builder, this move the computed
fields to the Report class. This functionality make more sense being in
the report class, as isn't to do with building a CSV. These changes are
pre-requisite work to simplify the CSV builder and use the Pipeline
class within the FileExporter class.