Add cleanup task to delete files in Google Drive
This adds a cleanup method for targets to delete any files that are 3
months or older. This is determined based on the createdAt time, which
should be accurate as the data within the file should corrospond to the
day that file was created.

To list files in a specific folder with Google Drive, you need to call
the `list_files` API method (behaves like a search method) where the
query (`q`) is defined as "show any files who have the parent folder x".
The page size is set 1000, so probably shouldn't need the use of
pagination as at most we should only see 90 files per 3 month period.