Support exporting an "event" SmartSurvey variable
We intend to allow this survey to be associated with different events
(such asking a particular minister a question). The approach to
distinguishing between submissions for these events (and non event
submissions) is intended to be via the use of a Smart Survey Custom
Variable [1].

This change provides support for a variable with a name of "event".
Later, we will adapt the export pipelines to include references to this
variable so it can be included in exports.

[1]: https://help.smartsurvey.co.uk/article/custom-variables