Include the event variable in Cabinet Office and third party exports
This adds an additional field to the CSV outputs for Cabinet Office and
third party exports of event. The event variable is an example of a
Smart Survey custom variable that is set by a query string (handling of
this was introduced in 9872334ad). This custom variable has been created
in the draft and live surveys.

The parties receiving this spreadsheet requested this be added as the
last column in their CSV to avoid this breaking any existing scripts
that process the sheets.